Move markdown extensions to separate module
